Move icon functionality into own file + add JSdocs + update README
