+++ src/js/main.js
@@ -7,7 +7,7 @@ import { getValue, mountEditor } from './spreadsheet/codeEditor.js';
 import { initDB, saveCellValue, getCellValue } from './spreadsheet/db.js';
 import consoleBtnsActiveState from './console/consoleBtns.mjs';
 import { showDropdownMenu } from './header/menu.mjs';
-import getIcon from './icons/index.js';
+import replaceIconsWithSVGs from './icons/replaceIconsWithSVGs.js';
 
 const spreadsheetContainer = document.querySelector('#spreadsheetContainer');
 
@@ -56,12 +56,4 @@ initDB()
     console.error('Failed to initialize IndexedDB:', error);
   });
 
-const iconSpans = document.querySelectorAll('[data-icon]');
-if (iconSpans) {
-  iconSpans.forEach((i) => {
-    const icon = getIcon(i.dataset.icon, i.dataset.size, i.dataset.class);
-    const parser = new DOMParser();
-    const svg = parser.parseFromString(icon, 'image/svg+xml');
-    i.replaceWith(svg.children[0]);
-  });
-}
+replaceIconsWithSVGs();
